Make PlantReader be a main class that will call StateReader and SequenceReader
